12.37.1.0

+++ com/meizu/cloud/pushsdk/notification/c/a.java
@@ -1,5 +1,6 @@
 package com.meizu.cloud.pushsdk.notification.c;
 
+import android.text.TextUtils;
 import com.baidu.searchbox.aperf.bosuploader.FileUtil;
 import com.meizu.cloud.pushinternal.DebugLogger;
 import java.io.File;
@@ -9,14 +10,14 @@ import java.io.FileOutputStream;
 /* loaded from: classes8.dex */
 public class a {
     public static void a(String str, String str2) {
+        String[] list;
         try {
             new File(str2).mkdirs();
-            String[] list = new File(str).list();
-            for (int i = 0; i < list.length; i++) {
-                File file = str.endsWith(File.separator) ? new File(str + list[i]) : new File(str + File.separator + list[i]);
+            for (String str3 : new File(str).list()) {
+                File file = str.endsWith(File.separator) ? new File(str + str3) : new File(str + File.separator + str3);
                 if (file.isFile()) {
                     FileInputStream fileInputStream = new FileInputStream(file);
-                    FileOutputStream fileOutputStream = new FileOutputStream(str2 + "/" + file.getName().toString());
+                    FileOutputStream fileOutputStream = new FileOutputStream(str2 + "/" + file.getName());
                     byte[] bArr = new byte[5120];
                     while (true) {
                         int read = fileInputStream.read(bArr);
@@ -29,7 +30,7 @@ public class a {
                     fileOutputStream.close();
                     fileInputStream.close();
                 } else if (file.isDirectory()) {
-                    a(str + "/" + list[i], str2 + "/" + list[i]);
+                    a(str + "/" + str3, str2 + "/" + str3);
                 }
             }
         } catch (Exception e) {
@@ -38,38 +39,44 @@ public class a {
     }
 
     public static boolean a(String str) {
-        File file = new File(str);
-        if (file.isFile() && file.exists()) {
-            return file.delete();
+        if (!TextUtils.isEmpty(str) && !str.contains("../")) {
+            File file = new File(str);
+            if (file.isFile() && file.exists()) {
+                return file.delete();
+            }
         }
         return false;
     }
 
     public static boolean b(String str) {
-        if (!str.endsWith(File.separator)) {
-            str = str + File.separator;
-        }
-        File file = new File(str);
-        if (file.exists() && file.isDirectory()) {
-            File[] listFiles = file.listFiles();
-            boolean z = true;
-            for (int i = 0; i < listFiles.length; i++) {
-                if (listFiles[i].isFile()) {
-                    z = a(listFiles[i].getAbsolutePath());
-                    if (!z) {
-                        break;
-                    }
-                } else {
-                    z = b(listFiles[i].getAbsolutePath());
-                    if (!z) {
-                        break;
+        File[] listFiles;
+        if (!TextUtils.isEmpty(str) && !str.contains("../")) {
+            if (!str.endsWith(File.separator)) {
+                str = str + File.separator;
+            }
+            File file = new File(str);
+            if (file.exists() && file.isDirectory()) {
+                boolean z = true;
+                for (File file2 : file.listFiles()) {
+                    boolean isFile = file2.isFile();
+                    String absolutePath = file2.getAbsolutePath();
+                    if (isFile) {
+                        z = a(absolutePath);
+                        if (!z) {
+                            break;
+                        }
+                    } else {
+                        z = b(absolutePath);
+                        if (!z) {
+                            break;
+                        }
                     }
                 }
+                if (z) {
+                    return file.delete();
+                }
+                return false;
             }
-            if (z) {
-                return file.delete();
-            }
-            return false;
         }
         return false;
     }
